tests: get rid of yield-based tests
yield-based tests (tests that are actually generating tests) are a nose
feature which were also supported in pytest. However, since pytest-3.0 this
type of tests is deprecated, they will be removed in pytest-4.0.

The only yield tests present in the kallithea test suite are simply yielding
a method that does a single assert. The benefit of having each of these
asserts in a separate test is not very high.

Replace the yield by a direct call to the previously-yielded method. This
means the total test count reduces but the actual amount of test code is the
same.