pullrequests: when creating, don't convert selected revision to tag name
It kind of makes sense that if a branch head or bookmarked revision is chosen,
then pretend the symbolic name was chosen. The name can move and will give the
PR a future. That is currently done both when populating the select boxes and
when actually creating the new PR. There, a selected revision is also
"upgraded" to its branch.

Tags are however mostly static. Fair enough if a user wants to create a PR from
a tag, but if a revision is selected, it shouldn't be side tracked to the tag name.