utils: when building a Mercurial ui object with configuration, don't use 'None' for NULL values
If Ui had an entry with
  ui_section='extensions', ui_key='largefiles', ui_value=NULL
it would be passed to Mercurial as if the .ini file had
  [extensions]
  largefiles = None
and it would fail to load the largefiles extension because it couldn't find
'./None/'.

Note: get_current_revision might currently mask this problem. It will not get
the Ui from the database and will thus read a normal .ini file from the system
and (if configured) read the largefiles extension from the default location.
That will make the largefiles extension available for later largefiles imports
even if they specify the bogus path.

As f8a714c2c5a1 noted in a FIXME: ui_value should perhaps not be nullable.

For now, just handle NULL in extension configuration.