helpers: remove direct global imports from model
To avoid circular imports, we need to determine some 'rules'.
Helpers have minimal dependencies to model. Most of the model imports that
are currently global are only used in a few methods. Therefore, it makes
sense to make these imports local to the method, so they won't 'count' for
circular imports at module level.