templates: fix aria markup from 24a0c176a63d
The markup in base.html is a bit misleading. The "quick login" drop down menu
is not just the quick login box but also the "current user info" box.

When the user is logged in, the "real" quick login box isn't shown and the drop
menu's aria-describedby="quick_login_h" thus becomes an invalid reference.

To fix the reference, also call the user info for quick_login_h - that is
approximately correct.