ini: set 'debug' only in one place
In Pylons-based Kallithea, the 'debug' option was set first in [DEFAULTS]
and then overridden with 'set debug = X' in [app:main]. Even when the value
under [DEFAULTS] was commented out, the use of the 'set' keyword (providing
override semantics) was required, because a default value for the global
'debug' was provided in the framework.

The presence of these two debug assignments is confusing.
Moreover, TurboGears2 makes the situation more easy and does not expect
'debug' to be in [DEFAULTS] nor does it provide a default at that level.
As a result, we can simple use 'debug = X' under [app:main].

Additionally, clarify the Big Fat Warning: the wording 'the line below' is
ambiguous and could be misunderstood in an ini file that had previously been
changed.