css: additional fixes
The fix in 56bce741176f was incorrect, as it robbed the login dialog of
its panel styling. Using on .panel-primary instead of #main solves this.

Additionally, while the rules deleted in a90acf3a1422 were indeed unused,
it didn't actually solve the problem of select2 dropdowns being white on
white, due to a separate selector also causing this effect. This should
be fixed now.